docs: Move example collection code to separate file
